checkout

+++ MeoCamp.Repository/Models/Order.cs
@@ -13,7 +13,7 @@ public partial class Order
 
     public DateTime? OrderDate { get; set; }
 
-    public decimal? TotalAmount { get; set; }
+    public int? TotalAmount { get; set; }
 
     public string OrderStatus { get; set; }
